requirements.txt
# Norwegian Aquaculture Regulatory Monitor
# Python 3.10+ required

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
html2text>=2024.2.26

# AI Summarization
anthropic>=0.40.0
openai>=1.50.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Email
sendgrid>=6.11.0

# Slack
slack-sdk>=3.27.0

# Dashboard
streamlit>=1.40.0
pandas>=2.2.0
plotly>=5.24.0

# Database
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
python-dateutil>=2.9.0
pytz>=2024.1
jinja2>=3.1.0

# Testing
pytest>=8.0.0
